fix db script

+++ f1api/src/main/java/com/f1api/repository/CircuitRepository.java
@@ -1,13 +1,12 @@
 package com.f1api.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import com.f1api.entity.Circuit;
 
 @Repository
-public interface CircuitRepository extends JpaRepository<Circuit, Short>{
+public interface CircuitRepository extends JpaRepository<Circuit, Integer>{
 
     
 }
